Squashed timeline architecture commit
This is a not-cleaned-up amalgamation of changes to separate the state
and behavior of game objects. The main benefits of this are:
- Easier to add interactions between objects. If one system needs the
  data from two types of objects (say, projectiles and players for
  collision detection), it doesn't need to ask them for that data.
  Instead, the data lives separately, and provides a full snapshot
  of game state.
- It removes one level of abstraction between the game state and things
  like network serialization and persistence (which could be useful in
  the future for replayability). We have the game's systems look at the
  data and construct the objects as needed, which is nice because we
  don't want to serialize objects over network. And it then makes no
  difference whether we start from persisted data vs live.

The "timeline" part of this change refers to us storing a timeline of
state snapshots per tick, which is very useful for reconciliation on the
client, and lag compensation on the server.