Apply "parse don't validate" to client simulation
Change 1:

Currently, game state on the client is stored as a flat dictionary
keyed by entity id. Each entity's data in turn is a dictionary. This
creates the issue that we have to repeatedly make nested dictionary
accesses and check the existence of any given piece of data we want.
For example, much of the current game logic assumes the player's own
character exists and has a physics state. In order to access this
data, we need to:
- Determine which entity is the player's character
- Look this up in the data map, ensuring it's not null
- Look up the physics state entry in the entity data map
It's messy to have all that code next to the actual game logic.
Furthermore, Godot's Dictionary type annotation isn't generic, so we
don't get any of the ergonomics/safety from that.

Let's solve this by storing client-side state as an object with typed
typed fields and accessors. We introduce some structure to this
object. Specifically, the distinction between the player's own
character and remote (other client) characters.

Change 2:

On the client, we keep the game state data separate from the objects
that calculate new states or display the current state. Decoupling
state and behavior is useful because sometimes we want to share state
across multiple game systems. Hypothetical example: predicting your own
character's movement might also require knowing the location of other
players for the purpose of collisions. Those locations also need to
be known by the logic for rendering other players. Rather than keeping
two copies of the same data in the objects that contain these two
pieces of logic, it's easier to keep it separate.

Anyhow, one downside of this is we need to make sure both the data
and the logic (really, the object doing some calculation) are available
at the time of the game tick. For example, if we know there is location
data for remote players, we also to guarantee there are player model
objects to actually display this data. Right now, these objects are
created and kept track of in a separate registry object. For each
tick, the client goes through the data and pulls the relevant objects
from the registry. This is done in the midst of running the actual
logic, which is bad because you could, for example, render some of
the remote players and then fail because another player doesn't
have their player model created yet.

You can think of this like trying to cook something complicated, and
needing certain ingredients at different stages of the dish. The
current approach is like being halfway through, realizing you should
add a lemon, and then scrambling to get the lemon but you remember
it isn't there. The dish is now a failure. Instead, we want to take
the approach of preparing all our ingredients beforehand. In the
game, this means we want to divide the logic into two stages:
1. Prepare all the data and logic objects we need
2. Execute the logic using the prepared resources

One benefit is that the execution code is much simpler because
it doesn't need to do nearly as many checks. There is still the risk
that the objects we need won't be ready even though we have the data.
However, we will now fail atomically, meaning we either run all the
logic successfully, or none at all. We will address this risk in the
future by directly creating the objects we need when we need them,
rather than managing their lifecycles separately.