Reduce redundant lookups and checks in server code
In order to run the game logic, the server needs a couple pieces of
information/behavior for each client:
1. The character data, i.e. physics state
2. The playermodel, first person camera, and physics body (we call these
   Game Objects)
3. A way to get the client's latest input, i.e. an input buffer

The server currently stores these in three separate places and has to
"manually" join together the data for a single client to run its logic.
Moreover, because the code to do this joining is scattered throughout
the business logic, we are forced to have null checks which can potentially
fail in the middle of executing a calculation.

To fix this, we take a similar approach as the client. Specifically, we
break the game logic into two broad steps:
1. Prepare all the resources (data + objects) we need
2. Execute the game logic, given these resources

Following the spirit of "parse don't validate," the output of step 1 is
structured so that it guarantees all the required inputs for step 2 are
present. Specifically, it maps from each client id to a struct which contains
exactly the required resources.

To guarantee input data will always be present, let's modify the
input buffer so that it returns a default value if empty, instead of
just returning a dummy value which causes us to skip the physics
calculation.

We also do a little hack to make it easier to "join" the related objects
for a client. We keep a map from client's ID's to a struct containing
the input buffer and the GameObjects needed to run logic. We choose
to still keep the data separate for now, as it makes little
conceptual sense for the state of the game to be scattered in many
different places (consider what would happen if we had maps like
this for projectiles, NPC's, etc.).