enable GDScript static type warnings
Static type annotations improve safety and make it more
ergonomic to program. We're already doing this by habit, but
let's enable the language feature to highlight cases we might
have missed and make it easier to follow the habit.

Not going to actually fix these everywhere in this commit,
but expect to do some cleanups as I work on features.