disable warn for static methods called on instance
We tend to mark methods as static whenever possible because it
conveys the valuable information that:
- It does not modify object state
- It does not depend on instance fields
It is convenient to access these without prefixing the class name
every time. Indeed, it feels unnatural to prefix when already in the
body of the class. GDScript warns about these calls by default.
Let's ignore them.