create entity objects directly
The current architecture separates the data for the entire game world
from the behavior. The data for all entities in the game is stored in a
central snapshot, which can basically be thought of as the entire state
of the world for a tick. The actual game logic and the fun parts (such
as being able to actually see the gameworld) live in objects (Godot
Scenes). For example, there's a collision body object which calculates
the player's movement, a player model object. etc. We call these objects
components. In order to progress and display the game world, we need
both the data and components for each entity.

Right now, the way we create these entities is via Godot's
MultiplayerSpawner. This boils down to calling a factory function on the
server, RPC'ing the client to do the same. The main problem here is that
the path by which we control object lifecycle, i.e. the RPC-ing done by
the spawner, is separate from the path where we send data to the client.
Ex: If we send snapshots to the client that contain the data for a
projectile, it's possible the spawner hasn't created the projectile
model yet. We're then forced to detect these situations and handle them
accordingly, which is not always easy to do in a graceful way (how do we
progress if our player isn't spawned?). Some of the work done in the
'parse don't validate' push makes this easier, but fundamentally doesn't
address the root cause.

A simpler solution to this problem is to just create the entities we
need based on what exists in the data. There's nothing stopping us from
just calling the factory function ourselves.

One caveat here is we now might be forced to do object creation and
adding those to the game world in the middle of our main loop
calculation. These might be slow, for example loading a big map. For
now, it's not an issue, but we can potentially address this by eagerly
spawning whatever we need, or adding some validation + default behavior
like we had before.