Separately measure input queue size at pop/push
Right now, we have one statistic that measures the size of the input
buffer. This is somewhat muddied because it samples the size both at
time of push, and time of pop. It is more informative to look at these
two separately. We are most worried about cases where we try to pop
when the queue is empty, or try to push when it's full. With separate
statistics,  we are guaranteed to know when this happens.