0: Add ability triggering on client
We separate the responsibilities of triggering an ability vs performing
it into two components. This will help us later when we want to do
different things in response to a trigger depending on if a character is
client's own player, remote player, or on server.

For now, action is a print statement on client side. We've re-used the
"slow walk" input for simplicity.