2: Extract base class for abilities
We want to be able to easily swap out a character's ability without
restructuring the CharacterComponents class. All the abilities in the
foreseeable future take the same handful of inputs, so we can just
make a mock "interface" via an abstract class and have abilities inherit
from that. CharacterComponents now has an AbstractAbilityAction, and
we can swap it for any implementation we want.