2: Add raycast ability to own client/server
For now, we aren't replicating to remote players on the client because
this is a little more involved. We'll need to make a way to transmit
triggers back to remote players, which means we'll have to think about
how to handle events (spooky).

We choose to implement our own sphere-ray intersect functionality
because Godot does not allow instantaneously setting the position
of a collision object and raycasting against the new position.
See https://github.com/godotengine/godot-proposals/issues/5181
This instant-raycasting is needed for lag compensation. An alternative
way to achieve this would be keeping a trailing tail of hitboxes, one
for each past frame. These could be on different physics layers, and
we raycast against the layer corresponding to the frame which the
player was seeing when they clicked. We only foresee needing sphere
colliders for this game, so it's simpler to roll our own vs doing the
extra book-keeping.