2: Replicate raycast for remote players
Triggering abilities should be replicated in a reliable way, unlike
movement which can afford to be interpolated/fudged. Consider how
annoying it would be to get hit by an ability you can't even see!

The state snapshots from server -> client are sent in an unreliable
way, because drops are generally recoverable and it's faster to use
unreliable delivery; we won't be held up if one packet keeps failing
to be delivered. So we need a separate route. Using an RPC for now
because it's simple.

In the future, we may want to formalize the idea of an "event," a more
general concept than just triggers. Events need to be sent reliably and
MUST be eventually by clients.