2: Expose camera transform calculation method
Currently, the first person display component's main display method
has two responsibilities:
1. Calculate the camera's position + basis
2. Move the camera to this transform
We are piggybacking off 1 to decide where our raycast points.

It's inconvenient that the transform calculation logic (which can
basically be static) is tied to the side effect of repositioning the
camera. Let's separate these. This also makes it cleaner to extract
the blocks of code that perform the raycast ability; they no longer
are forced to call the FPS display method again.