2: Extract top-level ability logic to method
Both the client and server have the code for running ability logic right
in the middle of all the other stuff (physics, display, etc.). We want
to eventually move all this logic into their own methods. Let's
start with the abilities, since there is a good amount of logic to
basically "ETL" the data into a form we can pass to the ability
component. This ETL stuff is irrelevant to other logic