3: Add bullet tracers
We use an immediate mesh because it's performant for rendering simple
shapes that get updated very frequently.

Initially, the idea was to have each ability component contain an object
which displays tracers. This works, but we think a better approach is
having the ability components return data about what they hit, and then
handle rendering the tracers separately. That way, it's very explicit
what the responsibility of the ability object is: It just calculates
"effects" on the game state, but the responsibility to apply and display
the effects lives elsewhere. One benefit is that if some other system
wants to know about the state of raycasts, it's much easier to do that
if their state isn't hidden within the ability component. For example,
a system to play an audio effect if a bullet passes near your player.