Break up server game logic into methods
We are currently running everything in physics process:
- Validate all inputs, data, components exist
- Parse the above into a map of "resources" per client id
- Calculate the next physics state
- Perform character abilities
- Apply bullet damage
- Arrange the data we need to export per client
- Display characters (first and third person)
- Display bullet tracers
- Send snapshots back to clients

Pretty much all the logic lives directly in the function body, which
although convenient has the downside of:
- visual clutter
- less clear exactly what each sub-system depends on, in terms of
  data, objects it must access, etc.

Let's break this logic up into logical parts + extract a static method
for each part so it's clear exactly what the inputs and outputs are.

Also, free-rider to this commit is we'll be changing how we format long
method signatures. We used to do this:
```
func bigMethod(
  input 1,
  input 2,
  ...
  ) -> return type:
  code 1
  code 2
```
Then we did this to have the code on different indent vs. the signature
```
func bigMethod(
    input 1,
    input 2,
    ...) -> return type:
  code 1
  code 2
```
On second thought, it does make sense to have the return type be on its
own line, because it's a piece of meaningful info. This is distinct
from method calls, where the closing parens adds no info and therefore
makes sense to just inline it
```
func bigMethod(
  input 1,
  input 2,
  ...
) -> return type:
  code 1
  code 2
```