Extricate debug drawing from hitscan ability
Right now, the hitscan ability class also draws debug spheres where
the rays hit. This caused a game breaking issue when we try to
despawn a character:
- Debug spheres are temporary, we start a timer that queue_frees
  the sphere after 5 seconds
- When we despawn a character, we free its components, including
  the ability component
- Debug sphere nodes are children of the ability component, and
  get freed along side it
- So if the sphere is freed, and then the timer tries to free it again,
  we get a null pointer

Couple ways to solve this, but we chose to just move the debug
drawing outside the ability; no reason to be in there. In general,
we want things that have state to live outside the components
because it makes it easier to reason about what they're doing.