6: reset client state on respawn
we rework the client-side respawning logic to be more explicit.
rather than relying on receiving a new state <-> respawned

this has the benefit that we can do all the respawn cleanup in one
place, and there is a clear death state (rather than death being
indistinguishable from extreme packet loss).

we keep the game simulation running for a few frames after
dying so that tracers will be displayed