Send redundant inputs to server
Due to packet loss and jitter, input messages might arrive late or
never at all. To guard against this, instead of only sending the input
for one tick in a message, we'll send the last N inputs. Now, even if
we drop a couple packets, it's very likely the next packet that arrives
will be able to "fill in" the information gap. This is quite cheap to
do because input messages are small in data size.

Tested this by measuring how often the server tries to pull an
input from the input buffer when the buffer is empty. Without the
change, we see around 500 misses per client in 30 seconds under
simulated 20% packet loss. With the change, we see 10 misses.

We're using a hardcoded message structure for now for simplicity.