Add counter statistics
Sometimes we just want to measure how many times something
occurs within a given time period. For example, it's useful to see
how many times the server tried to pop an input from an empty
client input buffer. We don't need the complexity of statistics for
this, so let's expose a simpler interface.

Right now, we're re-using the Statistics class for this for speed of
implementation, but feasibly it could be a different class.
We'll revisit this in the logging and metrics push later.