Extrapolate inputs when input buffer is empty
Even with the improved reliability we get from sending redundant
inputs in each message, there will be cases when we try to get
an input from the buffer when it is empty. We would always like
to have some input returned in these cases because otherwise it
creates a lot of complexity around how we should handle players
without an input.

Right now, we are returning a rather unreasonable default:
an input state with a 0,0 view angle, no movement, etc.
This is problematic because on frames where the buffer is empty,
the character pose appears to suddenly reset to the default angle
on the server and for remote characters on clients. Another major issue
here is it usually causes desync between the server and client's idea
of the character position. Generally, the dropped input packets would
have kept the player moving in the direction they were moving before,
but the current default causes the server to stop the player abruptly.

In the vast majority of cases, it is more sensible to re-use the last
known input, at least for a couple frames. Humans do not switch
inputs that rapidly or often, compared to the length of a few ticks.

We don't want to reuse the last input forever, in order to avoid
cases where players seem to walk indefinitely in a straight line
when network conditions are bad. After some point, we will
default to a standing-still input with the last known view angle.

We tested this by disabling redundant input messages and simulating a
10% packet rate loss. Even under these conditions, play felt smooth
from the client side, and we did not observe sudden pose switching on
server or remote clients.