Raise reconciliation max ticks
During our network stress testing, we found that at high latencies,
movement did not feel responsive on the client side. Why is this?
We observed "rubber-banding"-like behavior, which pointed at
faulty client-side reconciliation. Indeed, we saw that the client was
receiving authoritative states that were more than 16 ticks old,
which is the limit beyond which we will not re-simulate movement
for the purposes of reconciliation.

In plainer english: If the server tells us our player was at position X
100 ticks ago, we won't replay the last 100 ticks of input during
reconciliation. Instead, we only replay the last 16, which means the
client thinks that its true current state is much closer to position X
than the current position displayed on the client. The convergence
code then pulls the player back to X: This is the rubberbanding.

A very simple band-aid is to raise the limit from 16 -> 24.