Don't buffer own-player authoritative state
Currently, the world state snapshots sent by the server to the client
go into a buffer where they wait a few ticks before being consumed
by client logic. This helps protect against network jitter, but does
have the downside of adding some artificial latency.

This latency is generally fine, but isn't necessary or desirable for
authoritative state concerning the client's own character. Recall that
the client will locally predict player movement and use authoritative
state from the server to correct this prediction when it diverges from
the server's truth (reconciliation). There is no reason we need to
guard against network jitter for the purpose of reconciliation. It's
fine to just reconcile as soon as we get that information from the
server. Let's allow this data to skip the usual snapshot buffering, and
instead be used as soon as possible.