Make client state buffer tick-aware
Tick-aware means it will pop the snapshot with the lowest tick first,
and that it won't pop duplicate snapshots or stale snapshtos.

This is sort of "overdue" and makes sense ... we wouldnt' want to
display world state out of order just because the packets did not
arrive ordered. Using a buffer is a safeguard against network jitter.
This will also be helpful in the future when we implement message
redundancy and entity interpolation for state snapshots, as a guard
against packet loss. For example, for message redundancy, the
server might attach the last 3 states to each message, rather than
just the most recent one. When adding these to the buffer, it's nice
to know we are guaranteed not to double-up on already-received
states.