Add smooth entity interpolation on client
Under high packet-loss conditions, the client might not receive state
snapshots from the server for certain ticks. Right now, when we don't
have a new world state in the buffer, we just freeze the display
completely, which looks like all the remote characters flickering
between pausing for a few ticks and then moving again. This creates a
very stuttery experience for players.

As a kind of band-aid fix for this, instead of directly rendering the
latest authoritative state from our client buffer, we will instead lag
behind a bit using interpolation. That is, we will take:
- A: the last rendered state
- B: the next authoritative state from the buffer
And interpolate A to B by a fraction, displaying this interpolated
result.