Interpolate states in client snapshot buffer
Problem statement:
Let's say the server tries to send us snapshots for ticks 1, 2, and 3
Due to packet loss, snapshot 2 never reaches the client.
In the status quo, this means we'll never show snapshot 2, and there
will be a "jump" in the rendered world as we move instantly from
1 to 3. This is undesirable

Approach:
The client can take a reasonable guess of what snapshot 2 would have
been by simply interpolating halfway between 1 and 3. This won't
always be right, but it's much better than the current state, and is
also a battle-tested technique used in practice by many real
world multiplayer games.