Allow fully resetting snapshot buffer state
Currently, the client side death and respawn is handled by resetting
the state of the world, including the snapshot buffer. However, right
now we just clear the buffer and the last valid state, but there are
still parts of state we don't reset, like the variable tracking latest
received tick. Let's be more comprehensive and reset all state