Lag-compensate hitscan abilities
Problem:
Network latency causes the client to see the world state in the past
relative to the server. Therefore, when the client shoots at a moving
target and seems to hit it, the server might not agree because by
the time the client's shoot command reaches the server, the state of the
world on the server can be vastly different to what the client was
seeing at the time of the shot.

Example:
Assume one-way latency = 3 ticks == 50 ms
1. At time 0ms:
   - Client is displaying tick 12. Meaning it sees the state of the
      world corresponding to tick 12 on the server
   - Server is at tick 15
   - Client triggers a raycast against other player, which seems to hit
   - Client notifies server that it has done this raycast
2. At time 50ms
   - Server is at tick 18
   - Server gets the raycast notification from client
   - Server simulates the raycast again to check if the client hit
   - Sadly, the enemy player has moved a lot from tick 12 to tick 18,
     and the raycast misses.

Approach:
Let's use latency compensation to solve this problem. This boils down
to having the server simulate the raycast as though the current
server state were equal to what the client was seeing at the time it
took the shot. In practice, this just means we keep track of what each
player's position was on each server tick. The client tells us which
server tick it was displaying at the time of the shot, and then the
server calculates the raycast collisions using the cached positions
from that tick.

There is a nuance with determining which server tick the client was
displaying at the time of the shot, due to the entity state
interpolation that we previously added. This interpolation means that
the true rendered state lags behind a bit from the last state popped
from the state snapshot buffer. We estimate this lag amount as
(1 - entity_interp_fraction) / entity_interp_fraction
derived from some napkin math.

We tested how well the client's hitscan results (precise hit position)
lined up with the server before and after this change. With a simulated
latency of 200ms, the discrepancy goes from the scale of meters
to millimeters.

Good reading:
https://developer.valvesoftware.com/wiki/Lag_Compensation