Replicate ability triggers with precise timing
Problem:
When a client triggers an ability, other players should see
that they did so. The server, in other words, has to replicate
the trigger to all connected clients. Right now, we're doing
this by simply sending an RPC, which queues a "pending"
trigger for the character entity that triggered the ability.
On the next frame, we will compute and display the bullet
tracer based on the character's displayed transform on
that frame.

The issue with this is there's no guarantee that the world
state the client is currently seeing (i.e. the position and
rotation of the enemy players) is precisely the same as
the enemy player's pose at the time they fired the ability.

Approach:
Instead of this "Jesus take the wheel" style of replication,
let's embed a bit more information into this trigger, such as
the server tick at which it was triggered, and the exact position
and direction the player was facing when activating the ability.
We teach the client to use this information to display the ability
in sync with displaying the state of the world at the (server) tick
when the trigger happened. The extra info about position/rotation
is helpful to ensure the bullet tracers are at precisely the right
direction and origin.

Verification:
We observed that before this change, the debug spheres that are
drawn wherever bullets hit a player are drawn in very different
places on the server vs the remote replicated players. After
the change, it's very clear that these positions are much more
closely aligned, even at high latencies with moving players.