Add a menu for starting client/server
Right now, the main scene is optimized for local testing, it does:
- Try to start server on a specific ip / port
- If a server is already started there, start a client

This doesn't work for a game executable we can distribute to other
people. First off, there's  no way to choose the IP. Secondly, the
player doesn't get a choice whether to host or to connect. Third,
there's very little feedback on if a connection goes wrong.

Moreover, having a menu is a familiar pattern for launching a
game, and we'll need one eventually for things like settings.

The functionality added is:
- Ability to input an IP address (+ string validation)
- Choose to connect to client or server
- Timeout on trying to connect to client
- In-UI status text that tells you whether hosting/connecting was
  successful, and an error description if not

We also remove the logic in the Client class for connecting to
the server, because that's now handled in the main menu class.