Notify server of disconnect when client quits
Right now, when the client closes the game window, its connection to the
server is unceremoniously dropped. After some time, the server realizes
the client is no longer connected, and the disconnect cleanup kicks in.
It would be nicer to formally notify the server of a disconnect. Let's
add an on-quit hook to do this.

Implemented via autoload so that the behavior is guaranteed even if
we switch scenes.