Record inputs in client state timeline
Right now, we use the input handler to keep track of the client input
per tick. These are later used for movement reconciliation. There is
separately a state timeline that keeps a record of the world state per
client tick. Ideally, we want one source of truth on what the current
tick is. We are using the state timeline as that source of truth now,
but it's not very clean because we then have to ask it for the tick and
tell the input handler to record its latest input and associate it with
that tick. It would be more elegant  if there was just one tick-aware
timeline that keeps track of both state and input.

Giving it a bit of thought, input actually *is* part of the state, so
let's just put it into the state timeline. We can also apply this idea
to the server side to eliminate the "syncing" problem that we have
due to storing inputs separate from state.