Convert client state timeline to ringbuffer
We don't need to hold onto every state since the game start: This causes
memory usage to balloon. Instead, let's only retain the last 100 or so,
which should be sufficient.

In theory, there's a safety issue with retrieving items from the buffer
because we could return something that doesn't have the same tick
as what was requested (we only know the modulo is the same). This
probably isn't a big danger now because these old states are only
used for getting inputs for recon, and there is already a maximum
ticks in the past limit on how old of inputs we can retrieve.