Auto-start server if in headless mode
Right now, if the game is started in headless mode, it will do nothing.
Why? Because there's no way to push the start menu buttons. We only
plan to use headless mode for the dedicated server for now, so let's
skip the menu selection and just run the server in that case.