Extract network interactions from client/server
RPC definitions and references to the connection status are scattered
throughout client and server code.
1. Too much responsibility in Server and Client
2. Hard to test, because there's no clean way to mock network messages
   and events.

The server state snapshots and client inputs are fortunately passed
through a separate object (the NetworkMessenger), which does offload
some responsibility. Sadly, this messenger is not type aware, and just
passes a raw dictionary. This means serde logic is also in the Server
and Client classes.

Let's resolve these problems by extracting an "interface" which the
client and server talk to. This interface is an abstraction over all the
messages (game data) and events (network connect/disconnect) that
are relevant to the client or server.

Right now, this network bus is a frankenstein class, with no distinction
between client and server methods. We might extract that later, but
that's less return for cost, so not a prio now.