Simplify ServerCharacterState dataclass
This class does not need serde, and the getters add no value. They make
adding new fields more of a hassle. For now, let's just remove the
needless methods in this dataclass. In the future, we'd like to switch
to C# datamodels for easier enforcement of immutability, maybe
builder code generation, etc.