Move all simulation data to ServerCharacterState
Right now, ServerCharacterState only contains the physics and health
state. Information about the client's input is stored separately. This
forces us to match up input and character state whenever we have to
do a computation involving both, ex: physics or abilities. Not only is
this inconvenient, but also goes against the principle that all game
simulation state should be stored in one place, which is good for
simplicity in the long run.

This change - counterintuitively - increases the amount of code in
the server class, but it makes it easier to pull out the entire game
simulation logic (ie 60% of this class) in the future.