Extract gun model and animation to separate scene
It's simpler for the character display class to just tell the shotgun to
handle its own animation, rather than trying to use the animation tree.

The one downside of this is we won't get the gun model updated
automatically we export a new model from blender, but this is just
because export the gun and character together into the same Godot
scene. Not something worth addressing right now.