Use new character model for third-person display
Added a few pivot points to make animation simpler. Now, instead of
needing to set all the position and rotation information needed to
transform the raw imported model to the correct orientation,
each animation only needs to modify one attribute of a pivot point.
All the offsetting and orientating happens in a child of the pivot.

Example node structure:
- Pivot for head pitch (Node3d)
  - Pivot for head bobbing when walking
    - Head model (has a transform/rotation to face -Z)