Render tracers starting from gun barrel
Tracers currently look like they come out of the character's eye. Let's
instead make them originate from the gun barrel.

Let's do this by asking the gun model node what its position is, and
then updating hitscan results so they originate from that position.
The solution we use very much flies in the face of law of demeter,
but it gets the job done for now.

The next step will be pulling out a calculation stage which computes
the gun position (by applying a relative transform to camera position).
The output of this stage would then be used to position the gun model
and also to decide tracer origins. The benefit is we decouple tracer
display from gun model logic.