Add firing pattern and damage falloff
Non-atomic commit (my apolocheese)

Firing pattern = multiple bullets with fixed spread
We specify the pattern in angle space rather than position because it's
more intuitive. Going to hardcode the pattern for now, but it makes
sense to be a piece of plain data/configuration in the future.

Damage falloff:
- Below some distance, you do max damage
- Above some distance, you do min damage
- Between these, damage falloff is linear with distance

Simple as