Add client tick to InputState
It's easier to have the tick be part of the InputState datamodel because
they are always accessed together and almost always set together. The
one exception might be for extrapolated inputs on the server side, aka
when the server input buffer repeats the last received input in the case
where the buffer is empty from packet losses.