Add separate, reliable RPC for trigger inputs
When the player presses the button to fire their weapon, we want to
replicate this reliably to the server. However, normal input, such as
WASD movement, should be sent unreliably because it saves bandwidth
and reduces latency.

Right now, both trigger events and regular input are sent in the same
datamodel as part of an unreliable RPC. Let's make a separate route
for sending triggers. Another benefit is this simplifies the regular
input datamodel. It also made it easy to decouple handling triggers
from the usual input -> buffer pipeline, which means we can handle
them as soon as they arrive, reducing de-facto latency.

Later commit will remove now-unused code post this change.