Add byte serde to CharacterTransformState
The goal is reducing outbound server bandwidth / inbound client.
Currently, even with compression, the load is approx 10kB/s per client.
This increases O(n^2) with more clients. The downside of bigger packets
is generally higher latency and packet loss, plus bandwidth costs if we
hit limits for whatever cloud service we're using. When we tested on a
small DigitalOcean droplet, we nearly used 15% of available transfer
in just 20 minutes for this very reason.

Pull out serialization helpers into a utility class, and share logic
with InputState.