Tighten range for own position y axis serde
We were seeing slight up and down jittering after enabling byte
serde for state snapshots. This seemed to be an artefact with client
side reconciliation. Specifically, when resting on the floor, the
deserialized own-character-position we receive from the server would
be slightly different from the predicted position because we map the
y position to a 16-bit int, so it's not as granular as a double.

To fix, let's restrict the domain of the y position for serialization,
so that there is more granularity in the deserialized value. In the long
term, we want to separate own character state snapshots from the
state for the rest of the world, which will make it easier to send a
highly precise state without sacrificing bandwidth, because we can send
at a different cadence.