Add map directly instead of spawning it
Right now, we use a MultiplayerSpawner to replicate the map scene to
clients. This adds little value and has the drawback that connection
issues might lead the client to pre-emptively despawn the map, which
is jarring. Because there's only one possible map, let's just attach it
to the client/server scenes so it always exists during game sessions.