Play walking noise in first person
We use an animation tree as a convenient way to blend between silence
and walking at full speed. The blending ratio is based off a hardcoded
max speed for now.

Unlike the third-person display, clients have access to their own speed
and groundedness flag, so we can disable this sound when in air or when
only moving up and down.