Add version bump commit hook
We ran into a bug in a playtest where clients could not connect to
the server because they were on different versions. One was using
ZLIB compression on packets, the other wasn't. To make this easier
to catch in the future, let's add simple versioning to the game.

A simple way to do this is having a pre-commit hook write a random
string to a file. The game can print this string when it runs.