Add stats for rpc receive interval
We want to measure the interval between:
- Client inputs received on server
- State snapshots received by client

This captures network jitter + jitter in the send interval. The latter
is relevant because we call RPCs during the physics tick. This can
be problematic because the physics tick does not actually run at a
fixed interval. Rather, the engine has a semi-fixed time step. This
means that the physics steps actually run as a loop within the render
step, and the number of physics steps roughly depends on how long the
frame took. So for any given frame of time (i.e. around 16ms IRL) there
could be 0, 1, 2, or more physics frames. Typically, the RPC receiver
experiences this as "2 packets arrive in quick succession" or
"there is a long gap between 2 packets arriving."

Let's measure this so we can see how bad it is and know when we've
fixed it.