Spawn required game objects on demand
Much of the core server logic follows this pattern:
For some entity E, we need state S and some Godot node G in order to
do a computation. For example, we need a physics body and physics
state to compute next character position. We call the Godot nodes
"components."

Currently, the server keeps the state and object (behavior) for an
entity "together" by bundling them into a "StateAndComponent"
struct. The nice thing about this is it guarantees that if an entity has
a state, then it also has the required components. On the other hand,
it's clunky to work with this structure in 2 areas:
1. When saving/loading past versions of the state. It's hard to
   "persist" the objects alongside the state. For example, it would be
   very strange to have multiple character models of the same character
   on the screen. We want to keep copies of old states around for
   reconciliation and lag compensation.
2. It's one more layer of "wrapping" that we have to deal with. Would
  be simpler to just have a plain old data structure.
Not a deal breaker, but in the long term we foresee it being cleaner
to take an "ECS-like" approach, where we have some central data that
does not contain any behavior, and then systems which act, for all
intents and purposes, like stateless behavior. The systems will achieve
this by creating on demand any objects they need for calculations.

Let's take a step towards this by replacing StateAndComponents with
the get_or_create idiom, i.e. each system will create all the components
it needs on demand. Components for nonexistent entities will be
cleaned up at end of frame.